// src/app/modules/organization/organization-form/organization-form.enum.ts
export enum OrganizationFormFields {
  TYPE = 'type',
  CODE = 'code',
  NAME_AR = 'nameAr',
  NAME_EN = 'nameEn',
  PARENT_ORG = 'parentOrg',
  ENTITY_TYPE = 'entityType',
  LOGO = 'logo',
  IS_ACTIVE = 'isActive',
  OFFICIAL_EMAIL = 'officialEmail',
  MOBILE_NO = 'mobno',
  AREA = 'area',
  ADCDA_CLASSIFICATION = 'adcdaClassifcation',
  SECTOR = 'sector',
  ADCDA_PRIMARY = 'adcdaPrimary',
  INTERNAL = 'internal',
  USE_PARENT_LOGO = 'useParentLogo',
  LOGIN_UAE_PASS = 'loginUaePass',
  SYSTEM_USER = 'systemUser',
  LDAP_USER = 'ldapUser',
  LDAP_ORG_ID = 'ldapOrgId',
  LOGIN_OTP = 'loginOtp',
  LOGIN_CAPTCHA = 'loginCaptcha',
  CONTRACTOR_CONTRACT_NO = 'contractorContractNo',
  CONTRACTOR_EXPIRE_DATE = 'contractorExpDate',
  ORGANIZATION_CENTER = 'center',
  IS_LOCATION_SELECTED = 'isLocationSelcted',
  FEATURE_NAME = 'featureName',
  LOCATION = 'location'
}

export enum ENTITIY_TYPES_CODES {
  CONTRACTOR = 'CON',
  ORGANIZATION = 'ORG',
  Municipality_Center = 'MUCENTER',
  CENTER = 'CENTER',
  Division = 'DIV',
  Unit = 'UNIT',
  Point = 'POINT'
}